+++ src/layers/MultiscaleImageLayer/MultiscaleImageLayer.js
@@ -7,6 +7,9 @@ import MultiscaleImageLayerBase from './MultiscaleImageLayerBase';
 import ImageLayer from '../ImageLayer';
 import { to32BitFloat, onPointer } from '../utils';
 
+// From https://github.com/visgl/deck.gl/pull/4616/files#diff-4d6a2e500c0e79e12e562c4f1217dc80R128
+const DECK_GL_TILE_SIZE = 512;
+
 const defaultProps = {
   pickable: true,
   onHover: { type: 'function', value: null, compare: false },
@@ -23,7 +26,10 @@ const defaultProps = {
   lensBorderColor: { type: 'array', value: [255, 255, 255], compare: true },
   lensBorderRadius: { type: 'number', value: 0.02, compare: true },
   maxRequests: { type: 'number', value: 10, compare: true },
-  onClick: { type: 'function', value: null, compare: true }
+  onClick: { type: 'function', value: null, compare: true },
+  transparentColor: { type: 'array', value: null, compare: true },
+  refinementStrategy: { type: 'string', value: null, compare: true },
+  excludeBackground: { type: 'boolean', value: false, compare: true }
 };
 
 /**
@@ -39,16 +45,22 @@ const defaultProps = {
  * @param {Object} props.loader Loader to be used for fetching data.  It must implement/return `getTile`, `dtype`, `numLevels`, and `tileSize`, and `getRaster`.
  * @param {Array} props.loaderSelection Selection to be used for fetching data.
  * @param {String} props.id Unique identifier for this layer.
- * @param {String} props.onTileError Custom override for handle tile fetching errors.
- * @param {String} props.onHover Hook function from deck.gl to handle hover objects.
+ * @param {function} props.onTileError Custom override for handle tile fetching errors.
+ * @param {function} props.onHover Hook function from deck.gl to handle hover objects.
  * @param {boolean} props.isLensOn Whether or not to use the lens.
  * @param {number} props.lensSelection Numeric index of the channel to be focused on by the lens.
  * @param {number} props.lensRadius Pixel radius of the lens (default: 100).
- * @param {number} props.lensBorderColor RGB color of the border of the lens (default [255, 255, 255]).
+ * @param {Array} props.lensBorderColor RGB color of the border of the lens (default [255, 255, 255]).
  * @param {number} props.lensBorderRadius Percentage of the radius of the lens for a border (default 0.02).
  * @param {number} props.maxRequests Maximum parallel ongoing requests allowed before aborting.
- * @param {number} props.onClick Hook function from deck.gl to handle clicked-on objects.
- * @param {number} props.modelMatrix Math.gl Matrix4 object containing an affine transformation to be applied to the image.
+ * @param {function} props.onClick Hook function from deck.gl to handle clicked-on objects.
+ * @param {Object} props.modelMatrix Math.gl Matrix4 object containing an affine transformation to be applied to the image.
+ * @param {Array} props.transparentColor An RGB (0-255 range) color to be considered "transparent" if provided.
+ * In other words, any fragment shader output equal transparentColor (before applying opacity) will have opacity 0.
+ * This parameter only needs to be a truthy value when using colormaps because each colormap has its own transparent color that is calculated on the shader.
+ * Thus setting this to a truthy value (with a colormap set) indicates that the shader should make that color transparent.
+ * @param {string} props.refinementStrategy 'best-available' | 'no-overlap' | 'never' will be passed to TileLayer. A default will be chosen based on opacity.
+ * @param {boolean} props.excludeBackground Whether to exclude the background image. The background image is also excluded for opacity!=1.
  */
 
 export default class MultiscaleImageLayer extends CompositeLayer {
@@ -86,22 +98,26 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       lensBorderRadius,
       maxRequests,
       onClick,
-      modelMatrix
+      modelMatrix,
+      transparentColor,
+      excludeBackground,
+      onViewportLoad,
+      refinementStrategy
     } = this.props;
     const { tileSize, numLevels, dtype, isInterleaved, isRgb } = loader;
     const { unprojectLensBounds } = this.state;
+    // This is basically to invert:
+    // https://github.com/visgl/deck.gl/pull/4616/files#diff-4d6a2e500c0e79e12e562c4f1217dc80R128
+    // The z level can be wrong for showing the correct scales because of the calculation deck.gl does
+    // so we need to invert it for fetching tiles and minZoom/maxZoom.
+    const zoomOffset = Math.log2(DECK_GL_TILE_SIZE / tileSize);
     const noWebGl2 = !isWebGL2(this.context.gl);
     const getTileData = async ({ x, y, z, signal }) => {
       const tile = await loader.getTile({
         x,
         y,
-        // I don't fully undertstand why this works, but I have a sense.
-        // It's basically to cancel out:
-        // https://github.com/visgl/deck.gl/pull/4616/files#diff-4d6a2e500c0e79e12e562c4f1217dc80R128,
-        // which felt odd to me to beign with.
-        // The image-tile example works without, this but I have a feeling there is something
-        // going on with our pyramids and/or rendering that is different.
-        z: Math.round(-z + Math.log2(512 / tileSize)),
+        // See the above note within for why the use of zoomOffset and the rounding necessary.
+        z: Math.round(-z + zoomOffset),
         loaderSelection,
         signal
       });
@@ -125,11 +141,21 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       id: `Tiled-Image-${id}`,
       getTileData,
       dtype,
-      tileSize,
+      // If you scale a matrix up or down, that is like zooming in or out.  After
+      // https://github.com/visgl/deck.gl/pull/4616/files#diff-4d6a2e500c0e79e12e562c4f1217dc80R128,
+      // tileSize controls the zoom level that the tile indexer thinks you are at for fetching tiles.
+      // Because the indexing offsets `z` by math.log2(TILE_SIZE / tileSize), passing in
+      // tileSize * (1 / modelMatrix.getScale()[0]) from this layer as below to TileLayer gives an offset of
+      // math.log2(TILE_SIZE / (tileSize * (1 / modelMatrix.getScale()[0]))) = math.log2(TILE_SIZE / tileSize) + Math.log2(modelMatrix.getScale()[0])
+      // as desired so that the z level used for indexing the tiles is larger (i.e more zoomed in) if the image is scaled larger, and vice-versa if scaled smaller.
+      tileSize: modelMatrix
+        ? tileSize * (1 / modelMatrix.getScale()[0])
+        : tileSize,
       onClick,
       extent: [0, 0, width, height],
-      minZoom: -(numLevels - 1),
-      maxZoom: Math.min(0, Math.round(Math.log2(512 / tileSize))),
+      // See the above note within for why the use of zoomOffset and the rounding necessary.
+      minZoom: Math.round(-(numLevels - 1) + zoomOffset),
+      maxZoom: Math.round(zoomOffset),
       colorValues,
       sliderValues,
       channelIsOn,
@@ -137,7 +163,8 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       domain,
       // We want a no-overlap caching strategy with an opacity < 1 to prevent
       // multiple rendered sublayers (some of which have been cached) from overlapping
-      refinementStrategy: opacity === 1 ? 'best-available' : 'no-overlap',
+      refinementStrategy:
+        refinementStrategy || (opacity === 1 ? 'best-available' : 'no-overlap'),
       // TileLayer checks `changeFlags.updateTriggersChanged.getTileData` to see if tile cache
       // needs to be re-created. We want to trigger this behavior if the loader changes.
       // https://github.com/uber/deck.gl/blob/3f67ea6dfd09a4d74122f93903cb6b819dd88d52/modules/geo-layers/src/tile-layer/tile-layer.js#L50
@@ -155,7 +182,9 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       lensSelection,
       lensBorderColor,
       lensBorderRadius,
-      modelMatrix
+      modelMatrix,
+      transparentColor,
+      onViewportLoad
     });
     // This gives us a background image and also solves the current
     // minZoom funny business.  We don't use it for the background if we have an opacity
@@ -165,13 +194,17 @@ export default class MultiscaleImageLayer extends CompositeLayer {
     const layerModelMatrix = modelMatrix ? modelMatrix.clone() : new Matrix4();
     const baseLayer =
       implementsGetRaster &&
+      !excludeBackground &&
       new ImageLayer(this.props, {
         id: `Background-Image-${id}`,
         modelMatrix: layerModelMatrix.scale(2 ** (numLevels - 1)),
         visible:
-          opacity === 1 ||
-          (-numLevels > this.context.viewport.zoom &&
-            (!viewportId || this.context.viewport.id === viewportId)),
+          opacity === 1 &&
+          (!viewportId || this.context.viewport.id === viewportId) &&
+          // If we are using a transparent color, we shouldn't show the background image
+          // since the background image might not have the same color output from the fragment shader
+          // as the tiled layer at a higher resolution level.
+          !transparentColor,
         z: numLevels - 1,
         pickable: true,
         onHover,