+++ src/views/DetailView.js
@@ -1,6 +1,9 @@
-import { MultiscaleImageLayer, ImageLayer, ScaleBarLayer } from '../layers';
+import { ScaleBarLayer } from '../layers';
 import VivView from './VivView';
-import { getVivId } from './utils';
+import { getImageLayers, getVivId } from './utils';
+import { OVERVIEW_VIEW_ID } from './OverviewView';
+
+export const DETAIL_VIEW_ID = 'detail';
 
 /**
  * This class generates a MultiscaleImageLayer and a view for use in the VivViewer as a detailed view.
@@ -11,18 +14,7 @@ export default class DetailView extends VivView {
     const { loader } = props;
     const { id, height, width } = this;
     const layerViewState = viewStates[id];
-    const layers = [];
-
-    const detailLayer = loader.isPyramid
-      ? new MultiscaleImageLayer(props, {
-          id: `${loader.type}${getVivId(id)}`,
-          viewportId: id
-        })
-      : new ImageLayer(props, {
-          id: `${loader.type}${getVivId(id)}`,
-          viewportId: id
-        });
-    layers.push(detailLayer);
+    const layers = getImageLayers(id, props);
 
     const { physicalSizes } = loader;
     if (physicalSizes) {
@@ -42,5 +34,21 @@ export default class DetailView extends VivView {
     }
 
     return layers;
+  }
+
+  /**
+   * Create a viewState for this class, checking the id to make sure this class and viewState match.
+   * @param {Object} args
+   * @param {ViewState} args.ViewState ViewState object.
+   * @returns {ViewState} ViewState for this class (or null by default if the ids do not match).
+   */
+  filterViewState({ viewState, currentViewState }) {
+    if (viewState.id === OVERVIEW_VIEW_ID) {
+      const { target } = viewState;
+      if (target) {
+        return { ...currentViewState, target };
+      }
+    }
+    return super.filterViewState({ viewState });
   }
 }