+++ src/views/utils.js
@@ -1,4 +1,7 @@
 import { OrthographicView } from '@deck.gl/core';
+// Do not import from '../layers' because that causes a circular dependency.
+import MultiscaleImageLayer from '../layers/MultiscaleImageLayer';
+import ImageLayer from '../layers/ImageLayer';
 
 export function getVivId(id) {
   return `-#${id}#`;
@@ -25,32 +28,73 @@ export function makeBoundingBox(viewState) {
 }
 
 /**
- * Create an initial view state that centers the image in the view at the right zoom level.
- * @param {loader} Object The loader of the image for which the view state is desired.
- * @param {viewSize} Object { height, width } object for deducing the right zoom level to center the image.
+ * Create an initial view state that centers the image in the viewport at the zoom level that fills the dimensions in `viewSize`.
+ * @param {Object} loader The loader of the image for which the view state is desired.
+ * @param {Object} viewSize { height, width } object giving dimensions of the viewport for deducing the right zoom level to center the image.
+ * @param {Object} zoomBackOff A positive number which controls how far zoomed out the view state is from filling the entire viewport (default is 0 so the image fully fills the view).
+ * SideBySideViewer and PictureInPictureViewer use .5 when setting viewState automatically in their default behavior, so the viewport is slightly zoomed out from the image
+ * filling the whole screen.  1 unit of zoomBackOff (so a passed-in value of 1) corresponds to a 2x zooming out.
  * @returns {ViewState} A default initial view state that centers the image within the view: { target: [x, y, 0], zoom: -zoom }.
  */
-export function getDefaultInitialViewState(loader, viewSize) {
+export function getDefaultInitialViewState(loader, viewSize, zoomBackOff = 0) {
   const { height, width } = loader.getRasterSize({
     z: 0
   });
-  // Get a reasonable initial zoom level for pyramids based on screen size.
-  const { isPyramid } = loader;
-  let zoom = 0;
-  let size = Infinity;
-  // viewSize is not in the dependencies array becuase we only want to use it when the source changes.
-  if (isPyramid) {
-    while (size >= Math.max(...Object.values(viewSize))) {
-      const rasterSize = loader.getRasterSize({
-        z: zoom
-      });
-      size = Math.max(...Object.values(rasterSize));
-      zoom += 1;
-    }
-  }
+  const zoom =
+    Math.log2(Math.min(viewSize.width / width, viewSize.height / height)) -
+    zoomBackOff;
   const loaderInitialViewState = {
-    target: [height / 2, width / 2, 0],
-    zoom: isPyramid ? -zoom : -1.5
+    target: [width / 2, height / 2, 0],
+    zoom
   };
   return loaderInitialViewState;
+}
+
+/**
+ * Creates the layers for viewing an image in detail.
+ * @param {string} id The identifier of the view.
+ * @param {Object} props The layer properties.
+ * @returns {Array} An array of layers.
+ */
+export function getImageLayers(id, props) {
+  const {
+    loaderSelection,
+    newLoaderSelection,
+    onViewportLoad,
+    transitionFields,
+    ...layerProps
+  } = props;
+  const { loader } = layerProps;
+  // Create at least one layer even without loaderSelection so that the tests pass.
+  if (loader.isPyramid) {
+    return [loaderSelection, newLoaderSelection]
+      .filter((s, i) => i === 0 || s)
+      .map((s, i) => {
+        const suffix = s
+          ? `-${transitionFields.map(f => s[0][f]).join('-')}`
+          : '';
+        const newProps =
+          i !== 0
+            ? {
+                onViewportLoad,
+                refinementStrategy: 'never',
+                excludeBackground: true
+              }
+            : {};
+        return new MultiscaleImageLayer({
+          ...layerProps,
+          ...newProps,
+          loaderSelection: s,
+          id: `${loader.type}${getVivId(id)}${suffix}`,
+          viewportId: id
+        });
+      });
+  }
+  return [
+    new ImageLayer(layerProps, {
+      id: `${loader.type}${getVivId(id)}`,
+      viewportId: id,
+      loaderSelection
+    })
+  ];
 }